requirements.txt
dotmap>=1.2.20
flake8>=3.5.0
google-api-python-client>=1.6.5
h5py>=2.7.1
ipython>=6.2.1
jupyter_contrib_nbextensions>=0.4.0
jupyterlab>=0.31.12
keras>=2.1.4
imgaug>=0.2.5
matplotlib>=2.2.0
nbconvert>=5.3.1
nbformat>=4.4.0
nbstripout>=0.3.1
notebook>=5.4.0
numpy>=1.14.1
opencv-python>=3.4.3.18
Pillow>=5.0.0
pydicom>=1.0.2
scikit-image>=0.13.1
scipy>=1.0.0
tensorflow-tensorboard==1.5.1
tensorflow-gpu==1.5.0
yapf>=0.24.0